chore(Types): use intptr_t for casting pointers into integers
In ISO C/C++, there is no guarantee that pointers have the same width as `long`,
though ILP32 and LP64 do so.

* maple/xover.c:
    * Every_Z_Xover() param `arg`
* so/adminutil.c:
    * m_resetsys() param `arg`

+++ maple/xover.c
@@ -2228,7 +2228,7 @@ Every_Z_Favorite(void)
 static int
 Every_Z_Xover(const void *arg)
 {
-    xover((int)(unsigned int)(long)arg);
+    xover((int)(unsigned int)(intptr_t)arg);
     return 0;
 }
